@@ -1560,14 +1560,12 @@ public class JParser {
           .collect(Collectors.toList());
 
         TypeTree patternType = convertType(recordPattern.getPatternType());
-        IdentifierTreeImpl recordName = recordPattern.getPatternName() != null ? convertSimpleName(recordPattern.getPatternName()) : null;
-        return new RecordPatternTreeImpl(patternType, nestedPatterns, recordName);
+        return new RecordPatternTreeImpl(patternType, nestedPatterns);
       case ASTNode.GUARDED_PATTERN:
         GuardedPattern g = (GuardedPattern) p;
         return new GuardedPatternTreeImpl(
           convertPattern(g.getPattern()),
-          // FIXME java 19 support: should be "TerminalTokens.TokenNameRestrictedIdentifierWhen" instead of "TerminalTokens.TokenNameIdentifier"
-          firstTokenBefore(g.getExpression(), TerminalTokens.TokenNameIdentifier),
+          firstTokenBefore(g.getExpression(), TerminalTokens.TokenNameRestrictedIdentifierWhen),
           convertExpression(g.getExpression()));
       case ASTNode.NULL_PATTERN:
         // It is not clear how to reach this one, it seems to be possible only with badly constructed AST
@@ -1780,12 +1778,7 @@ public class JParser {
   private YieldStatementTreeImpl convertYield(YieldStatement e) {
     InternalSyntaxToken yieldKeyword = null;
     if (!e.isImplicit()) {
-      try {
-        yieldKeyword = firstTokenIn(e, TerminalTokens.TokenNameRestrictedIdentifierYield);
-      } catch (AssertionError | IndexOutOfBoundsException error) {
-        // TODO ECJ bug? should be "TerminalTokens.TokenNameRestrictedIdentifierYield" in all cases
-        yieldKeyword = firstTokenIn(e, TerminalTokens.TokenNameIdentifier);
-      }
+      yieldKeyword = firstTokenIn(e, TerminalTokens.TokenNameRestrictedIdentifierYield);
     }
     return new YieldStatementTreeImpl(
       yieldKeyword,
@@ -2263,8 +2256,7 @@ public class JParser {
   private InstanceOfTreeImpl convertInstanceOf(PatternInstanceofExpression e) {
     Expression leftOperand = e.getLeftOperand();
     InternalSyntaxToken instanceofToken = firstTokenAfter(leftOperand, TerminalTokens.TokenNameinstanceof);
-    //FIXME future versions of ECJ are likely to return a Pattern rather than a SingleVariableDeclaration
-    return new InstanceOfTreeImpl(convertExpression(leftOperand), instanceofToken, new TypePatternTreeImpl(convertVariable(e.getRightOperand())));
+    return new InstanceOfTreeImpl(convertExpression(leftOperand), instanceofToken, convertPattern(e.getPattern()));
   }
 
   private LambdaExpressionTreeImpl convertLambdaExpression(LambdaExpression e) {
