@@ -47,6 +47,8 @@ class JavaVersionImplTest {
     assertThat(version.isJava17Compatible()).isFalse();
     assertThat(version.isJava18Compatible()).isFalse();
     assertThat(version.isJava19Compatible()).isFalse();
+    assertThat(version.isJava20Compatible()).isFalse();
+    assertThat(version.isJava21Compatible()).isFalse();
     assertThat(version.asInt()).isEqualTo(-1);
   }
 
@@ -68,6 +70,9 @@ class JavaVersionImplTest {
     assertThat(version.isJava17Compatible()).isEqualTo(javaVersionAsInt >= 17);
     assertThat(version.isJava18Compatible()).isEqualTo(javaVersionAsInt >= 18);
     assertThat(version.isJava19Compatible()).isEqualTo(javaVersionAsInt >= 19);
+    assertThat(version.isJava20Compatible()).isEqualTo(javaVersionAsInt >= 20);
+    assertThat(version.isJava21Compatible()).isEqualTo(javaVersionAsInt >= 21);
+
     assertThat(version.asInt()).isEqualTo(javaVersionAsInt);
   }
 
@@ -87,9 +92,9 @@ class JavaVersionImplTest {
 
   @Test
   void test_effective_java_version() {
-    assertThat(new JavaVersionImpl().effectiveJavaVersionAsString()).isEqualTo("19");
+    assertThat(new JavaVersionImpl().effectiveJavaVersionAsString()).isEqualTo("21");
     assertThat(new JavaVersionImpl(10).effectiveJavaVersionAsString()).isEqualTo("10");
-    assertThat(new JavaVersionImpl(-1).effectiveJavaVersionAsString()).isEqualTo("19");
+    assertThat(new JavaVersionImpl(-1).effectiveJavaVersionAsString()).isEqualTo("21");
   }
 
   @Test
